% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validar_rut.R
\name{validar_rut}
\alias{validar_rut}
\title{Validar formato de un RUT}
\usage{
validar_rut(rut)
}
\arguments{
\item{rut}{Vector de RUTs a consultar}
}
\value{
TRUE/FALSE
}
\description{
Aplica varias pruebas para confirmar que un RUT es correcto.
}
\details{
Para confirmar que un RUT es válido, detecta que contenga un guión, verifica el formato "xxx-x", y valida que el dígito verificador sea correcto. Esta función acepta uno o varios RUTs.
}
