% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_digito.R
\name{calcular_digito}
\alias{calcular_digito}
\title{Calcular dígito verificador del RUT}
\usage{
calcular_digito(rut_sin_digito)
}
\arguments{
\item{rut_sin_digito}{Vector de RUTs sin dígito verificador}
}
\value{
Dígito verificador del cuerpo del RUT entregrado, en tipo caracter
}
\description{
Calcula el dígito verificador para un RUT sin dígito verificador. Puede recibir uno o varios RUTs.
}
