% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_rut.R
\name{limpiar_rut}
\alias{limpiar_rut}
\title{Limpiar RUTs}
\usage{
limpiar_rut(ruts)
}
\arguments{
\item{ruts}{Vector de texto con RUTs}
}
\value{
Vector de texto con RUTs limpios en formato \code{xxxxxxx-y}
}
\description{
Recibe un vector de posibles RUT, y los limpia eliminando puntuación y espacios, extrayendo números, y separando el dígito verificador
}
\examples{
rut_sucios <- c(
  "24.444.145-9",
  "24444145 9",
  "24 444 145 9",
  "24,444,145,9",
  "2M4A4P4A4C1H4E59",
  "244441459",
  "hola hola")

limpiar_rut(rut_sucios)
}
